Catch exceptions at this point
If an exception is thrown from here then the exception will likely propagate all the way up to the top of pnpm,
yielding a scary-looking stack trace that looks like pnpm crashed. Ideally if the user makes an error in their package.json (such as a space instead of a hyphen in the package name), they get an error message that points them towards their error.